refactor: rename app.api.domain to app.api.schema
