+++ app/api/adapter/character.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from app.api.domain.character import CharacterResponse
+from app.api.schema.character import CharacterResponse
 from app.domain.character import Character
 from app.services.character import (service_get_character_by_id, service_list_characters,
                                     service_search_characters_by_name, service_create_character,
