+++ app/api/schema/character.py
