+++ app/api/schema/character_attribute.py
