+++ app/api/schema/character_class.py
